+++ api-user/src/main/java/com/user/security/jwt/JwtFilter.java
@@ -1,5 +1,8 @@
-package com.user.security;
+package com.user.security.jwt;
 
+import com.user.security.CustomUserDetailService;
+import com.user.security.CustomUserDetails;
+import com.user.service.TokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
