+++ api-user/src/main/java/com/user/security/jwt/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package com.user.security;
+package com.user.security.jwt;
 
 import com.storage.enums.TokenType;
 import io.jsonwebtoken.JwtException;
