+++ api-user/src/main/java/com/user/service/AuthService.java
@@ -11,7 +11,6 @@ import com.storage.user.UserRepository;
 import com.user.global.exception.ConflictException;
 import com.user.global.exception.ErrorCode;
 import com.user.global.exception.NotFoundException;
-import com.user.security.TokenService;
 import com.user.web.request.ReissueTokenRequest;
 import com.user.web.request.UserSignInRequest;
 import com.user.web.request.UserSignUpRequest;
